[AIP-21] Move contrib.operators.file_to_wasb providers.microsoft.azure.operators.file_to_wasb
