+++ scripts/ci/pylint_todo.txt
@@ -39,7 +39,7 @@
 ./airflow/providers/amazon/aws/operators/emr_add_steps.py
 ./airflow/providers/amazon/aws/operators/emr_create_job_flow.py
 ./airflow/providers/amazon/aws/operators/emr_terminate_job_flow.py
-./airflow/contrib/operators/file_to_wasb.py
+./airflow/providers/microsoft/azure/operators/file_to_wasb.py
 ./airflow/providers/grpc/operators/grpc.py
 ./airflow/providers/jenkins/operators/jenkins_job_trigger.py
 ./airflow/providers/jira/operators/jira.py
